Some Wrap-Up Completed    - Solved #63 issue: Hide Keyboard When Exit a Fragment    - Completed some JavaDoc

+++ app/src/main/java/ibrahim/example/stocklogger/fragments/SellStockFragment.java
@@ -30,6 +30,7 @@ import java.util.Calendar;
 
 import ibrahim.example.stocklogger.R;
 import ibrahim.example.stocklogger.databases.StockDatabase;
+import ibrahim.example.stocklogger.pojos.ActiveStock;
 import ibrahim.example.stocklogger.pojos.SoldStock;
 import ibrahim.example.stocklogger.pojos.Stock;
 
@@ -140,37 +141,75 @@ public class SellStockFragment extends Fragment  {
                                 .setAction("Action", null).show();
                     }
 
-                    if(!error){
+                    if(soldQuantityEditText.getText().toString().equals("")){
+                        error = true;
+                        YoYo.with(Techniques.Shake)
+                                .duration(700)
+                                .playOn(soldQuantityEditText);
+                        Snackbar.make(view, "Sold quantity can't be blank", Snackbar.LENGTH_LONG)
+                                .setAction("Action", null).show();
+                    }
 
+                    if(!error){
                         double soldPrice = Double.parseDouble(soldPriceEditText.getText().toString());
+                        int soldQuantity = Integer.parseInt(soldQuantityEditText.getText().toString());
                         double worth = stock.getWorth();
                         double rating = 1;
 
-                        if(stock.isUSD()){
-                            SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(getContext());
-                            rating = Double.parseDouble(sp.getString("isUSD","1.26"));
+
+                        // If sold part of a stock
+                        if(soldQuantity<stock.getQuantity()){
+                            // get trading fee
+                            double tradingFee = 6.95;
+                            SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(getContext());
+                            tradingFee = Double.parseDouble(sharedPreferences.getString("tradingFee", "6.95"));
+
+                            // add a new record to active stock
+                            StockDatabase db = new StockDatabase(getContext());
+                            ActiveStock activeStock = new ActiveStock(
+                                    stock.getSymbol(),
+                                    stock.getCompanyName(),
+                                    soldPrice,
+                                    soldQuantity * -1,
+                                    calendarTextDate.getText().toString()
+                            );
+                            int activeStockId = db.addActiveStock(activeStock);
+
+                            // Update stock worth and quantity
+                            double soldWorth = (activeStock.getPrice() * activeStock.getQuantity()+ tradingFee) / activeStock.getQuantity();
+                            stock.setWorth(worth - soldWorth);
+                            stock.setQuantity(stock.getQuantity() - soldQuantity);
+                            db.updateStock(stock);
+
+                            // add a new relationship to stock_active table
+                            db.addStockActive(stock.getId(), activeStockId);
+                            db.close();
+                        } else {
+                            // If sold all holding stocks of a company
+                            if (stock.isUSD()) {
+                                SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(getContext());
+                                rating = Double.parseDouble(sp.getString("isUSD", "1.26"));
+                            }
+
+
+                            double earned = (soldPrice - worth) * soldQuantity * rating;
+
+                            StockDatabase db = new StockDatabase(getContext());
+
+                            db.addSoldStock(
+                                    new SoldStock(
+                                            stock.getSymbol(),
+                                            stock.getCompanyName(),
+                                            soldPrice,
+                                            earned,
+                                            calendarTextDate.getText().toString()
+                                    )
+                            );
+
+                            db.deleteStock(stock.getId());
+
+                            db.close();
                         }
-
-                        int quantity = Integer.parseInt(soldQuantityEditText.getText().toString());
-
-                        double earned = (soldPrice - worth) * quantity * rating;
-
-                        StockDatabase db = new StockDatabase(getContext());
-
-                        db.addSoldStock(
-                                new SoldStock(
-                                        stock.getSymbol(),
-                                        stock.getCompanyName(),
-                                        soldPrice,
-                                        earned,
-                                        calendarTextDate.getText().toString()
-                                )
-                        );
-
-                        db.deleteStock(stock.getId());
-
-                        db.close();
-
                         Navigation.findNavController(view).popBackStack();
                     }
                 }
